refactor: extract cli option parsing into module
