+++ atch_cli_opts.h
@@ -1,0 +1,6 @@
+#ifndef ATCH_CLI_OPTS_H
+#define ATCH_CLI_OPTS_H
+
+int atch_parse_options(int *argc, char ***argv);
+
+#endif
