+++ findings.md
@@ -7,6 +7,7 @@
 - Introducing `libatch.a` before `atch` in Make changed default target implicitly; fixed via `.DEFAULT_GOAL := atch`.
 - Session name/path expansion was a good first reusable extraction target because it had clear boundaries and minimal side effects.
 - Command alias matching/dispatch is another clean extraction point that reduces CLI frontend coupling.
+- Option parsing is a third good extraction target because it is self-contained and shared by multiple command paths.
 
 ## Decisions
 - Keep core behavior unchanged; avoid moving large code blocks in first refactor.
@@ -14,5 +15,6 @@
 - Treat this as phase 1+2 internal modularization toward future library split.
 
 ## Notes
-- `atch_expand_session_name_dup` now centralizes bare-name expansion and cache-dir creation behavior for future frontends.
+- `atch_expand_session_name_dup` centralizes bare-name expansion and cache-dir creation behavior for future frontends.
 - `atch_resolve_command` centralizes alias policy (`list/l/ls`, `attach/a`, etc.) for reuse outside the CLI binary.
+- `atch_parse_options` now centralizes option semantics and validation, reducing direct coupling in `atch.c`.
